+++ frontend/src/app/data-entry/create-forms/create-entity/create-entity.graphql
