+++ frontend/src/app/data-entry/create-forms/create-entity/create-space.service.ts
