+++ frontend/src/app/data-entry/create-forms/create-forms.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
-import { CreateAgentComponent } from './create-agent/create-agent.component';
+import { CreateAgentComponent } from './create-entity/create-entity.component';
 import { SharedModule } from '@shared/shared.module';
 import { DataEntrySharedModule } from '../shared/data-entry-shared.module';
-import { CreateAgentService } from './create-agent/create-agent.service';
-import { CreateGiftService } from './create-agent/create-gift.service';
-import { CreateLetterService } from './create-agent/create-letter.service';
-import { CreateSpaceService } from './create-agent/create-space.service';
-import { CreateEntityService } from './create-agent/create-entity.service';
+import { CreateAgentService } from './create-entity/create-agent.service';
+import { CreateGiftService } from './create-entity/create-gift.service';
+import { CreateLetterService } from './create-entity/create-letter.service';
+import { CreateSpaceService } from './create-entity/create-space.service';
+import { CreateEntityService } from './create-entity/create-entity.service';
