+++ frontend/src/app/data-entry/episode-form/episode-entities-form/episode-entities-form.component.html
@@ -44,6 +44,6 @@
 
         <lc-create-agent [entityType]="entityType" [create]="createEntity$"
             [sourceID]="data.episode.source.id"
-            [episodeID]="data.episode.id" />
+            [episodeID]="data.episode.id" />ya
     </ng-container>
 </ng-container>
